'''Misc'''
 * Bandits and Genetic Programming 
 * threadOwnedExecutionContext / SubExecutionContext: fix pushWorkUnit() -- the ExecutionStack was not correct 
 * OptimizerState: fix a small bug on autosave 
 * PopulationBasedOptimizer: make the bests variables unique 
 * MaximumEntropyClassifier: uniform distribution when the parameters are not learned yet 
 * MaximumEntropyClassifier: implement clone()
 * EnumerationSampler: fix the return type of sample()

+++ include/lbcpp/Optimizer/OptimizerState.h
@@ -86,7 +86,7 @@ extern ClassPtr optimizerStateClass;
 class DistributionBasedOptimizerState : public OptimizerState
 {
 public:
-  DistributionBasedOptimizerState(size_t autoSaveStateFrequency = 0) : OptimizerState(autoSaveStateFrequency) {}
+  DistributionBasedOptimizerState(double autoSaveStateFrequency = 0) : OptimizerState(autoSaveStateFrequency) {}
   
   const DistributionPtr& getDistribution() const
     {return distribution;}
@@ -105,7 +105,7 @@ typedef ReferenceCountedObjectPtr<DistributionBasedOptimizerState> DistributionB
 class SamplerBasedOptimizerState : public OptimizerState
 {
 public:
-  SamplerBasedOptimizerState(const SamplerPtr& sampler, size_t autoSaveStateFrequency = 0)
+  SamplerBasedOptimizerState(const SamplerPtr& sampler, double autoSaveStateFrequency = 0)
     : OptimizerState(autoSaveStateFrequency), sampler(sampler), initialSampler(sampler) {}
   SamplerBasedOptimizerState() {}
